refactor(core): extract facade test owners
